changing order structure

+++ app/pages/FileManagement.vue
@@ -117,6 +117,9 @@ async function createFolder(folder: string) {
 }
 
 async function deleteFolder(item: string) {
+  if (!confirm("Sind Sie sicher, dass sie diesen Ordner löschen möchten?"))
+    return;
+
   try {
     await $fetch("/api/files", {
       method: "DELETE",
@@ -125,6 +128,33 @@ async function deleteFolder(item: string) {
         item: item,
       },
     });
+
+    load(currentPath.value);
+  } catch (e: unknown) {
+    toast.add({
+      title: "Fehler",
+      description:
+        "Beim Löschen des Ordners ist ein Fehler aufgetreten! Wenden Sie sich an Ihren Administrator!",
+      color: "error",
+      icon: "i-heroicons-x-mark",
+    });
+  }
+}
+
+async function changeFolder(item: string) {
+  if (!confirm("Sind Sie sicher, dass sie diesen Ordner umbenennen möchten?"))
+    return;
+
+  try {
+    // await $fetch("/api/files", {
+    //   method: "DELETE",
+    //   body: {
+    //     path: currentPath.value,
+    //     item: item,
+    //   },
+    // });
+
+    console.log(item);
 
     load(currentPath.value);
   } catch (e: unknown) {
@@ -176,10 +206,10 @@ onMounted(() => {
       <li
         v-for="i in items"
         :key="i.name"
-        class="cursor-pointer hover:bg-gray-100 rounded px-2 py-1 transition flex items-center gap-1 text-xl"
+        class="hover:bg-gray-100 rounded px-2 py-1 transition flex items-center gap-3 text-xl"
       >
-        <div @click="open(i)">
-          <div v-if="i.type === 'dir'" class="flex gap-2">
+        <div class="cursor-pointer" @click="open(i)">
+          <div v-if="i.type === 'dir'" class="flex gap-2 items-center">
             <svg
               xmlns="http://www.w3.org/2000/svg"
               fill="none"
@@ -196,7 +226,7 @@ onMounted(() => {
             </svg>
             {{ i.name }}
           </div>
-          <div v-else class="flex gap-2">
+          <div v-else>
             <svg
               xmlns="http://www.w3.org/2000/svg"
               fill="none"
@@ -215,7 +245,7 @@ onMounted(() => {
           </div>
         </div>
 
-        <div class="hover:text-red-500" @click="deleteFolder(i.name)">
+        <button class="hover:text-red-500" @click="deleteFolder(i.name)">
           <svg
             xmlns="http://www.w3.org/2000/svg"
             fill="none"
@@ -230,7 +260,24 @@ onMounted(() => {
               d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0"
             />
           </svg>
-        </div>
+        </button>
+
+        <button class="hover:text-yellow-500" @click="changeFolder(i.name)">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            fill="none"
+            viewBox="0 0 24 24"
+            stroke-width="1.5"
+            stroke="currentColor"
+            class="size-6"
+          >
+            <path
+              stroke-linecap="round"
+              stroke-linejoin="round"
+              d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L6.832 19.82a4.5 4.5 0 0 1-1.897 1.13l-2.685.8.8-2.685a4.5 4.5 0 0 1 1.13-1.897L16.863 4.487Zm0 0L19.5 7.125"
+            />
+          </svg>
+        </button>
       </li>
 
       <br />
